% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_vs_bayes_incl_bar.R
\name{bayes_perc_cat}
\alias{bayes_perc_cat}
\title{Percentage of edges in the respective evidence categories}
\usage{
bayes_perc_cat(agg_point)
}
\arguments{
\item{agg_point}{A data frame with the edge specific information.}
}
\value{
A bar chart that shows the percentage of edges in the respective categories.
}
\description{
This function creates a bar chart that shows the percentage of edges falling into respective edge categories.
}
