% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge_est_vs_post_incl_prob.R
\name{edge_est_vs_post_incl_prob}
\alias{edge_est_vs_post_incl_prob}
\title{Network Edge Estimates Vs. Posterior Inclusion Probabilities}
\usage{
edge_est_vs_post_incl_prob(agg_point)
}
\arguments{
\item{agg_point}{A data frame with the estimates for each edge, for all studies.}
}
\value{
A scatter plot that compares the network edge estimates to their posterior inclusion probabilities.
}
\description{
This function creates a scatter plot that compares the network edge estimates to their posterior inclusion probabilities.
}
