% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge_evidence_plot.R
\name{edge_evidence_plot}
\alias{edge_evidence_plot}
\title{Network Edge Inclusion Evidence Plot}
\usage{
edge_evidence_plot(network_results, inclusion = TRUE)
}
\arguments{
\item{network_results}{A list containing the results from fitting the network model.}

\item{inclusion}{A logical value indicating whether to plot the edge inclusion evidence (if FALSE, then evidence for exclusion is plotted).}
}
\value{
A network plot of the parameters from the BGGM results.
}
\description{
This function creates a network plot of the evidence for edge inclusion from the BGGM results.
}
