% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_metadata.R
\name{extract_metadata}
\alias{extract_metadata}
\title{Relevant Network Metadata from `AggStudyResults.RData`}
\usage{
extract_metadata(entry)
}
\arguments{
\item{entry}{A list entry from the `AggStudyResults.RData` file.}
}
\value{
A data frame with the following columns:
\itemize{
  n_nodes: The number of nodes in the network.
  n_edges: The number of edges in the network.
  sample_size: The sample size of the study.
  model: The model used to generate the network.
  topic: The topic of the study.
  subtopic: The subtopic of the study.}
}
\description{
This function extracts relevant metadata from the `AggStudyResults.RData` file.
}
