% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_vs_bayes_est_plot.R
\name{freq_vs_bayes_est_plot}
\alias{freq_vs_bayes_est_plot}
\title{Frequentist vs. Bayesian Parameter Estimate Scatterplot}
\usage{
freq_vs_bayes_est_plot(agg_point)
}
\arguments{
\item{agg_point}{A data frame with the estimates for each edge, for all studies.}
}
\value{
A scatterplot of the frequentist and Bayesian parameter estimates.
}
\description{
This function creates a scatterplot of the frequentist and Bayesian parameter estimates.
}
