% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_density_plot.R
\name{net_density_plot}
\alias{net_density_plot}
\title{Network Edge Density}
\usage{
net_density_plot(agg_point)
}
\arguments{
\item{agg_point}{A data frame with the estimates for each edge, for all studies.}
}
\value{
A density plot of the proportion of included edges in all networks.
}
\description{
This function creates a density plot of the proportion of included edges in all networks.
}
