% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_par_plot.R
\name{network_par_plot}
\alias{network_par_plot}
\title{Network Parameter Plot}
\usage{
network_par_plot(network_results)
}
\arguments{
\item{network_results}{A list containing the results from fitting the network model.}
}
\value{
A network plot of the parameters from the BGGM results.
}
\description{
This function creates a network plot of the parameters from the BGGM results.
}
