% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_citation.R
\name{parse_citation}
\alias{parse_citation}
\title{Parse Citation into In-text Reference, Year and DOI}
\usage{
parse_citation(citation)
}
\arguments{
\item{citation}{A character vector containing the citation in APA format.}
}
\value{
A list containing the in-text reference, year and DOI.
}
\description{
This function extracts the in-text reference, year and DOI from a citation.
}
