% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_vs_bayes_incl_bar.R
\name{freq_vs_bayes_incl_bar}
\alias{freq_vs_bayes_incl_bar}
\title{Frequentist Vs. Bayesian Edge Inclusion Bar Chart}
\usage{
freq_vs_bayes_incl_bar(agg_level)
}
\arguments{
\item{agg_level}{A data frame with the aggregated evidence for inclusion for each edge.}
}
\value{
A bar chart that compares the Bayesian and Frequentist edge inclusions.
}
\description{
This function creates a bar chart that compares the Bayesian and Frequentist edge inclusions,
}
